[CSL-2406] Create Haskell utility to generate X.509 certificates for TLS authentication
This is a pure Haskell implementation (no OpenSSL) to be used during installation to generate certificates now required
for authenticating Daedalus with Cardano-SL (see CBR-82). A couple of things are configurable on certificates, such as
the Subject Alternative Names for the server or the expiration date (default to 10 years for CA and 1 for other certs).

By default, the utility doesn't output the CA private key (which is therefore lost after having signed all certificates).